requirements.txt
absl-py==1.4.0
anvil==0.0.2
anvil-uplink==0.4.2
appnope==0.1.3
array-record==0.4.1
asttokens==2.4.0
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cycler==0.12.0
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
dm-tree==0.1.8
etils==1.5.2
executing==2.0.0
filelock==3.12.4
Flask==3.0.0
fonttools==4.43.0
fsspec==2023.10.0
future==0.18.3
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
importlib-resources==6.1.1
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
keras==2.15.0
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.0
oauth2client==4.1.3
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydot==2.0.0
PyDrive==1.3.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
sklearn==0.0.post9
stack-data==0.6.3
tensorflow-datasets==4.9.3
tensorflow-metadata==1.14.0
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.24.5
wcwidth==0.2.8
Werkzeug==3.0.1
wrapt==1.16.0
ws4py==0.5.1
zipp==3.17.0

tensorflow~=2.14.0
